requirements.txt
flask
jinja2
asn1crypto==1.3.0
autopep8==1.5
cffi==1.14.0
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.0.0
pycodestyle==2.5.0